refactor(speedtest-tracker): address all PR review feedback
- Add speedtestTrackerEnvelopeSchema to eliminate 'as' casts in integration
- Extract private getAsync helper mirroring homeassistant pattern
- Add comment on 404 = no tests run yet (fresh install)
- Bump cache TTL from 30s to 15 minutes
- Extract sections into separate component files (helpers, section-label, speed-stat-card, latest-result, averages, recent-results)
- Add comment explaining why recharts tick prop requires custom SVG for multi-line labels
