+++ packages/widgets/src/speedtest-tracker/component.tsx
@@ -1,24 +1,18 @@
 "use client";
 
 import { useMemo } from "react";
-import { AreaChart, ChartTooltip } from "@mantine/charts";
-import { Card, Flex, Group, SimpleGrid, Stack, Text, useMantineTheme } from "@mantine/core";
-import { useElementSize } from "@mantine/hooks";
-import { IconArrowDown, IconArrowUp, IconCircleCheck, IconCircleX, IconWaveSine } from "@tabler/icons-react";
-import { ReferenceLine, XAxis } from "recharts";
+import { Stack, Text } from "@mantine/core";
 
 import { clientApi } from "@homarr/api/client";
-import { useRequiredBoard } from "@homarr/boards/context";
-import type {
-  SpeedtestTrackerDashboardData,
-  SpeedtestTrackerResult,
-  SpeedtestTrackerStats,
-} from "@homarr/integrations/types";
+import type { SpeedtestTrackerDashboardData } from "@homarr/integrations/types";
 import { useScopedI18n } from "@homarr/translation/client";
-import type { TablerIcon } from "@homarr/ui";
 
 import type { WidgetComponentProps } from "../definition";
 import { NoIntegrationDataError } from "../errors/no-data-integration";
+import { AveragesSection } from "./averages";
+import { mergeStats, parseTimestamp } from "./helpers";
+import { LatestResultSection } from "./latest-result";
+import { RecentResultsSection } from "./recent-results";
 
 export default function SpeedtestTrackerWidget({
   options,
@@ -117,409 +111,3 @@ function SpeedtestTrackerContent({ integrationIds, options, isEditMode }: Speedt
     </Stack>
   );
 }
-
-// ─── Sections ─────────────────────────────────────────────────────────────────
-
-function LatestResultSection({ result }: { result: SpeedtestTrackerResult }) {
-  const t = useScopedI18n("widget.speedtestTracker");
-  const timestamp = parseTimestamp(result.created_at).toLocaleString([], {
-    month: "short",
-    day: "numeric",
-    hour: "2-digit",
-    minute: "2-digit",
-  });
-  const cols = result.healthy !== null ? 4 : 3;
-
-  return (
-    <Stack gap={6} h="100%">
-      <Group justify="space-between" align="center" wrap="nowrap">
-        <SectionLabel>{t("latestResult")}</SectionLabel>
-        <Text size="xs" c="dimmed">
-          {timestamp}
-        </Text>
-      </Group>
-      <SimpleGrid cols={cols} spacing="xs" style={{ flex: 1, gridTemplateRows: "1fr" }}>
-        <SpeedStatCard
-          icon={IconArrowDown}
-          color="blue"
-          value={formatResultSpeed(result, "download")}
-          label="Download"
-        />
-        <SpeedStatCard icon={IconArrowUp} color="teal" value={formatResultSpeed(result, "upload")} label="Upload" />
-        <SpeedStatCard
-          icon={IconWaveSine}
-          color="orange"
-          value={result.ping !== null ? `${result.ping.toFixed(1)} ms` : "—"}
-          label="Ping"
-        />
-        {result.healthy !== null && (
-          <SpeedStatCard
-            icon={result.healthy ? IconCircleCheck : IconCircleX}
-            color={result.healthy ? "green" : "red"}
-            value={result.healthy ? t("healthy") : t("unhealthy")}
-            label="Status"
-          />
-        )}
-      </SimpleGrid>
-    </Stack>
-  );
-}
-
-function AveragesSection({ stats }: { stats: SpeedtestTrackerStats }) {
-  const t = useScopedI18n("widget.speedtestTracker");
-
-  return (
-    <Stack gap={6} h="100%">
-      <SectionLabel>
-        {t("averages")} · {stats.total_results} {t("tests")}
-      </SectionLabel>
-      <SimpleGrid cols={3} spacing="xs" style={{ flex: 1, gridTemplateRows: "1fr" }}>
-        <SpeedStatCard
-          icon={IconArrowDown}
-          color="blue"
-          value={formatStatsSpeed(stats.download)}
-          label="Avg Download"
-          compact
-        />
-        <SpeedStatCard
-          icon={IconArrowUp}
-          color="teal"
-          value={formatStatsSpeed(stats.upload)}
-          label="Avg Upload"
-          compact
-        />
-        <SpeedStatCard
-          icon={IconWaveSine}
-          color="orange"
-          value={`${stats.ping.avg.toFixed(1)} ms`}
-          label="Avg Ping"
-          compact
-        />
-      </SimpleGrid>
-    </Stack>
-  );
-}
-
-function RecentResultsSection({ results }: { results: SpeedtestTrackerResult[] }) {
-  const t = useScopedI18n("widget.speedtestTracker");
-  const { ref, height } = useElementSize<HTMLDivElement>();
-
-  return (
-    <Stack gap={4} h="100%" style={{ minHeight: 0 }}>
-      <SectionLabel>{t("recentResults")}</SectionLabel>
-      <div ref={ref} style={{ flex: 1, minHeight: 0 }}>
-        {height > 0 && <SpeedHistoryChart results={results} height={height} />}
-      </div>
-    </Stack>
-  );
-}
-
-// ─── Speed history chart ──────────────────────────────────────────────────────
-
-function SpeedHistoryChart({ results, height }: { results: SpeedtestTrackerResult[]; height: number }) {
-  const board = useRequiredBoard();
-  const theme = useMantineTheme();
-  const data = useMemo(
-    () =>
-      [...results]
-        .sort(
-          (resultA, resultB) =>
-            parseTimestamp(resultA.created_at).getTime() - parseTimestamp(resultB.created_at).getTime(),
-        )
-        // Skip failed/zero-value results — treat 0 bits as "no result"
-        .filter((result) => (result.download_bits ?? 0) > 0)
-        .map((result) => ({
-          ts: parseTimestamp(result.created_at).getTime(),
-          Download: parseFloat(((result.download_bits ?? 0) / 1_000_000).toFixed(2)),
-          Upload:
-            result.upload_bits != null && result.upload_bits > 0
-              ? parseFloat((result.upload_bits / 1_000_000).toFixed(2))
-              : 0,
-        })),
-    [results],
-  );
-
-  // ── Y axis: nice round ticks ──────────────────────────────────────────────
-  const yConfig = useMemo(() => {
-    const maxVal = Math.max(...data.map((datum) => Math.max(datum.Download, datum.Upload)), 0);
-    const step = maxVal <= 400 ? 100 : 200;
-    const roundedMax = Math.ceil(maxVal / step) * step || step;
-    const ticks: number[] = [];
-    for (let val = 0; val <= roundedMax; val += step) ticks.push(val);
-    return { ticks, domain: [0, roundedMax] as [number, number] };
-  }, [data]);
-
-  // ── X axis: 2-hour interval ticks + midnight ──────────────────────────────
-  const { midnightTs, xTicks, topDateTicks } = useMemo(() => {
-    if (data.length === 0) return { midnightTs: null, xTicks: [] as number[], topDateTicks: [] as number[] };
-
-    const first = data[0]?.ts ?? 0;
-    const last = data.at(-1)?.ts ?? 0;
-
-    // 2-hour interval ticks aligned to wall-clock hours
-    const twoHourMs = 2 * 60 * 60 * 1000;
-    const startTick = Math.ceil(first / twoHourMs) * twoHourMs;
-    const tickSet = new Set<number>();
-    for (let t = startTick; t <= last; t += twoHourMs) tickSet.add(t);
-
-    // Find local midnight within the data window
-    const midnightDate = new Date(first);
-    midnightDate.setHours(24, 0, 0, 0);
-    const mTs = midnightDate.getTime();
-    const foundMidnight = mTs > first && mTs < last ? mTs : null;
-    if (foundMidnight) tickSet.add(foundMidnight);
-
-    // Top date axis: first data point + any midnight boundaries
-    const topTicks: number[] = [first];
-    if (foundMidnight) topTicks.push(foundMidnight);
-
-    return {
-      midnightTs: foundMidnight,
-      xTicks: Array.from(tickSet).sort((tsA, tsB) => tsA - tsB),
-      topDateTicks: topTicks,
-    };
-  }, [data]);
-
-  // Custom tick: time for all ticks, date label below time at midnight
-  const renderXTick = ({ x, y, payload }: { x: number; y: number; payload: { value: number } }) => {
-    const timestamp = payload.value;
-    const date = new Date(timestamp);
-    const timeStr = date.toLocaleTimeString([], { hour: "2-digit", minute: "2-digit" });
-    const isMidnight = midnightTs !== null && timestamp === midnightTs;
-    const dateStr = isMidnight ? date.toLocaleDateString([], { month: "short", day: "numeric" }) : null;
-
-    return (
-      <g transform={`translate(${x},${y})`}>
-        <text x={0} y={12} textAnchor="middle" fontSize={10} fill="var(--mantine-color-dimmed)">
-          {timeStr}
-        </text>
-        {dateStr && (
-          <text x={0} y={23} textAnchor="middle" fontSize={9} fill="var(--mantine-color-dimmed)">
-            {dateStr}
-          </text>
-        )}
-      </g>
-    );
-  };
-
-  // Top date axis tick: shows short date label above the chart
-  const renderTopDateTick = (props: { x: number; y: number; payload: { value: number }; index: number }) => {
-    const { x, y, payload, index } = props;
-    const date = new Date(payload.value);
-    const dateStr = date.toLocaleDateString([], { weekday: "short", month: "short", day: "numeric" });
-    const anchor = index === 0 ? "start" : "middle";
-    return (
-      <g transform={`translate(${x},${y})`}>
-        <text x={0} y={-4} textAnchor={anchor} fontSize={10} fill="var(--mantine-color-dimmed)">
-          {dateStr}
-        </text>
-      </g>
-    );
-  };
-
-  if (data.length === 0) return null;
-
-  return (
-    <AreaChart
-      h={height}
-      data={data}
-      dataKey="ts"
-      series={[
-        { name: "Download", color: "blue" },
-        { name: "Upload", color: "teal" },
-      ]}
-      curveType="monotone"
-      gridAxis="y"
-      tickLine="none"
-      withLegend
-      fillOpacity={0.2}
-      styles={{ root: { padding: 5, borderRadius: theme.radius[board.itemRadius] } }}
-      valueFormatter={(val: number) => `${val} Mbps`}
-      xAxisProps={{
-        type: "number",
-        domain: ["dataMin", "dataMax"],
-        ticks: xTicks,
-        tick: renderXTick,
-        interval: 0,
-        // Extra height for date label below time at midnight ticks
-        height: midnightTs !== null ? 38 : 28,
-      }}
-      yAxisProps={{
-        ticks: yConfig.ticks,
-        domain: yConfig.domain,
-        tickFormatter: (val: number) => `${val}`,
-        width: 50,
-        tick: { fontSize: 10 },
-      }}
-      tooltipProps={{
-        // eslint-disable-next-line @typescript-eslint/no-explicit-any
-        content: (props: any) => {
-          const { label, payload, active } = props as {
-            label?: number | string;
-            payload?: Record<string, unknown>[];
-            active?: boolean;
-          };
-          if (!active || !payload?.length) return null;
-          const timestamp = typeof label === "number" ? label : Number(label);
-          const dateStr = Number.isNaN(timestamp)
-            ? ""
-            : new Date(timestamp).toLocaleString([], {
-                month: "short",
-                day: "numeric",
-                hour: "2-digit",
-                minute: "2-digit",
-              });
-          return <ChartTooltip label={dateStr} payload={payload} valueFormatter={(val: number) => `${val} Mbps`} />;
-        },
-      }}
-    >
-      {/* Top date axis — shows day labels above the chart */}
-      <XAxis
-        xAxisId="topDates"
-        orientation="top"
-        type="number"
-        dataKey="ts"
-        domain={["dataMin", "dataMax"]}
-        ticks={topDateTicks}
-        tick={renderTopDateTick}
-        height={20}
-        axisLine={false}
-        tickLine={false}
-        interval={0}
-      />
-      {/* Dotted vertical reference lines at each x-axis tick */}
-      {xTicks.map((tickTs) => (
-        <ReferenceLine
-          key={tickTs}
-          x={tickTs}
-          yAxisId="left"
-          stroke="var(--mantine-color-dimmed)"
-          strokeDasharray="3 3"
-          strokeOpacity={0.35}
-        />
-      ))}
-    </AreaChart>
-  );
-}
-
-interface SpeedStatCardProps {
-  icon: TablerIcon;
-  color: string;
-  value: string;
-  label: string;
-  compact?: boolean;
-}
-
-function SpeedStatCard({ icon: Icon, color, value, label, compact = false }: SpeedStatCardProps) {
-  const { ref, height, width } = useElementSize<HTMLDivElement>();
-  const board = useRequiredBoard();
-  const isWide = width > height + 20;
-  const hideLabel = height > 0 && height <= 38;
-
-  return (
-    <Card
-      ref={ref}
-      p={compact ? "xs" : "sm"}
-      radius={board.itemRadius}
-      bg={`var(--mantine-color-${color}-light)`}
-      h="100%"
-      style={{ flex: 1 }}
-    >
-      <Flex
-        h="100%"
-        w="100%"
-        align="center"
-        justify="center"
-        direction={isWide ? "row" : "column"}
-        gap={isWide ? 8 : 4}
-      >
-        <Icon size={compact ? 16 : 20} color={`var(--mantine-color-${color}-5)`} style={{ flexShrink: 0 }} />
-        <Flex direction="column" align={isWide ? "flex-start" : "center"} gap={0}>
-          <Text size={compact ? "sm" : "md"} fw={700} ta="center" lh={1.1}>
-            {value}
-          </Text>
-          {!hideLabel && (
-            <Text size="xs" c="dimmed" ta="center" lh={1.3}>
-              {label}
-            </Text>
-          )}
-        </Flex>
-      </Flex>
-    </Card>
-  );
-}
-
-// ─── Small helpers ────────────────────────────────────────────────────────────
-
-function SectionLabel({ children }: { children: React.ReactNode }) {
-  return (
-    <Text size="xs" fw={600} c="dimmed" tt="uppercase" style={{ letterSpacing: "0.05em" }}>
-      {children}
-    </Text>
-  );
-}
-
-// ─── Pure data helpers ────────────────────────────────────────────────────────
-
-/**
- * Parse a Speedtest Tracker timestamp. The API returns UTC timestamps without
- * a timezone indicator (e.g. "2026-03-28 05:45:00"). We append "Z" to force
- * UTC parsing so JS converts it to local time correctly.
- */
-const parseTimestamp = (timestamp: string): Date => new Date(`${timestamp.replace(" ", "T")}Z`);
-
-const mergeStats = (
-  statsA: SpeedtestTrackerDashboardData["stats"],
-  statsB: SpeedtestTrackerDashboardData["stats"],
-): SpeedtestTrackerDashboardData["stats"] => {
-  if (!statsB) return statsA;
-  if (!statsA) return statsB;
-  return {
-    ping: {
-      avg: (statsA.ping.avg + statsB.ping.avg) / 2,
-      min: Math.min(statsA.ping.min, statsB.ping.min),
-      max: Math.max(statsA.ping.max, statsB.ping.max),
-    },
-    download: {
-      avg: (statsA.download.avg + statsB.download.avg) / 2,
-      avg_bits:
-        statsA.download.avg_bits !== undefined && statsB.download.avg_bits !== undefined
-          ? (statsA.download.avg_bits + statsB.download.avg_bits) / 2
-          : (statsA.download.avg_bits ?? statsB.download.avg_bits),
-      min: Math.min(statsA.download.min, statsB.download.min),
-      max: Math.max(statsA.download.max, statsB.download.max),
-    },
-    upload: {
-      avg: (statsA.upload.avg + statsB.upload.avg) / 2,
-      avg_bits:
-        statsA.upload.avg_bits !== undefined && statsB.upload.avg_bits !== undefined
-          ? (statsA.upload.avg_bits + statsB.upload.avg_bits) / 2
-          : (statsA.upload.avg_bits ?? statsB.upload.avg_bits),
-      min: Math.min(statsA.upload.min, statsB.upload.min),
-      max: Math.max(statsA.upload.max, statsB.upload.max),
-    },
-    total_results: statsA.total_results + statsB.total_results,
-  };
-};
-
-const formatBitsPerSec = (bps: number): string => {
-  if (bps >= 1_000_000_000) return `${(bps / 1_000_000_000).toFixed(2)} Gbps`;
-  if (bps >= 1_000_000) return `${(bps / 1_000_000).toFixed(2)} Mbps`;
-  if (bps >= 1_000) return `${(bps / 1_000).toFixed(2)} Kbps`;
-  return `${bps} bps`;
-};
-
-const formatResultSpeed = (result: SpeedtestTrackerResult, dir: "download" | "upload"): string => {
-  const human = dir === "download" ? result.download_bits_human : result.upload_bits_human;
-  if (human) return human;
-  const bits = dir === "download" ? result.download_bits : result.upload_bits;
-  if (bits != null) return formatBitsPerSec(bits);
-  return "—";
-};
-
-const formatStatsSpeed = (band: SpeedtestTrackerStats["download"]): string => {
-  if (band.avg_bits_human) return band.avg_bits_human;
-  if (band.avg_bits != null) return formatBitsPerSec(band.avg_bits);
-  // avg is bytes/s — convert to bits/s
-  return formatBitsPerSec(band.avg * 8);
-};
